Default method resolution.
Most specific interface implementation wins over empty class. While class is declared on the same level as interface B it does not provide an implementation of the conflicting method and its implementation is taken from interface A. So resolution takes place between A and B where B is more specific. Thus, implementation from B is used.